added implicit animations sample page
