+++ Microsoft.Toolkit.Uwp.SampleApp/SamplePages/Implicit Animations/ImplicitAnimationsPage.xaml
@@ -1,0 +1,33 @@
+﻿<Page
+    x:Class="Microsoft.Toolkit.Uwp.SampleApp.SamplePages.ImplicitAnimationsPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:extensions="using:Microsoft.Toolkit.Uwp.UI.Extensions"
+    xmlns:animations="using:Microsoft.Toolkit.Uwp.UI.Animations"
+    mc:Ignorable="d">
+
+    
+    <!-- Need to have an placeholder of each element so they can be used in the XAML editor -->
+    <Grid extensions:VisualExtensions.CenterPoint="50,50,0">
+        <animations:Implicit.ShowAnimations>
+            <animations:TranslationAnimation></animations:TranslationAnimation>
+            <animations:OpacityAnimation></animations:OpacityAnimation>
+            <animations:RotationAnimation></animations:RotationAnimation>
+            <animations:RotationInDegreesAnimation></animations:RotationInDegreesAnimation>
+            <animations:OffsetAnimation></animations:OffsetAnimation>
+            <animations:ScalarAnimation></animations:ScalarAnimation>
+            <animations:ScaleAnimation></animations:ScaleAnimation>
+            <animations:Vector2Animation>
+                <animations:Vector2KeyFrame></animations:Vector2KeyFrame>
+                <animations:Vector3KeyFrame></animations:Vector3KeyFrame>
+                <animations:Vector4KeyFrame></animations:Vector4KeyFrame>
+                <animations:ExpressionKeyFrame></animations:ExpressionKeyFrame>
+                <animations:ScalarKeyFrame></animations:ScalarKeyFrame>
+            </animations:Vector2Animation>
+            <animations:Vector3Animation></animations:Vector3Animation>
+            <animations:Vector4Animation></animations:Vector4Animation>
+        </animations:Implicit.ShowAnimations>
+    </Grid>
+</Page>
