--- Microsoft.Toolkit.Uwp.SampleApp/SamplePages/ImplicitAnimations/ImplicitAnimationsPage.xaml
@@ -1,105 +1,0 @@
-﻿<Page
-    x:Class="Microsoft.Toolkit.Uwp.SampleApp.SamplePages.ImplicitAnimationsPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Microsoft.Toolkit.Uwp.SampleApp.SamplePages"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:extensions="using:Microsoft.Toolkit.Uwp.UI.Extensions"
-    xmlns:animations="using:Microsoft.Toolkit.Uwp.UI.Animations"
-    mc:Ignorable="d">
-
-    <Page.Resources>
-        <Style x:Key="CustomButtonStyle" TargetType="Button">
-            <Setter Property="Background" Value="{ThemeResource ButtonBackground}"/>
-            <Setter Property="Foreground" Value="{ThemeResource ButtonForeground}"/>
-            <Setter Property="BorderBrush" Value="{ThemeResource ButtonBorderBrush}"/>
-            <Setter Property="BorderThickness" Value="{ThemeResource ButtonBorderThemeThickness}"/>
-            <Setter Property="Padding" Value="8,4,8,4"/>
-            <Setter Property="HorizontalAlignment" Value="Left"/>
-            <Setter Property="VerticalAlignment" Value="Center"/>
-            <Setter Property="FontFamily" Value="{ThemeResource ContentControlThemeFontFamily}"/>
-            <Setter Property="FontWeight" Value="Normal"/>
-            <Setter Property="FontSize" Value="{ThemeResource ControlContentThemeFontSize}"/>
-            <Setter Property="UseSystemFocusVisuals" Value="True"/>
-            <Setter Property="FocusVisualMargin" Value="-3"/>
-            <Setter Property="Template">
-                <Setter.Value>
-                    <ControlTemplate TargetType="Button">
-                        <Grid x:Name="RootGrid" Background="{TemplateBinding Background}" extensions:VisualExtensions.KeepCenterPointCentered="True">
-                            <VisualStateManager.VisualStateGroups>
-                                <VisualStateGroup x:Name="CommonStates">
-                                    <VisualState x:Name="Normal">
-                                        <VisualState.Setters>
-                                            <Setter Target="RootGrid.(extensions:VisualExtensions.Scale)" Value="1"></Setter>
-                                        </VisualState.Setters>
-                                    </VisualState>
-                                    <VisualState x:Name="PointerOver">
-                                        <VisualState.Setters>
-                                            <Setter Target="RootGrid.(extensions:VisualExtensions.Scale)" Value="2"></Setter>
-                                        </VisualState.Setters>
-                                    </VisualState>
-                                    <VisualState x:Name="Pressed">
-                                    </VisualState>
-                                    <VisualState x:Name="Disabled">
-                                    </VisualState>
-                                </VisualStateGroup>
-                            </VisualStateManager.VisualStateGroups>
-                            <ContentPresenter x:Name="ContentPresenter" AutomationProperties.AccessibilityView="Raw"
-                                              BorderThickness="{TemplateBinding BorderThickness}" BorderBrush="{TemplateBinding BorderBrush}" ContentTemplate="{TemplateBinding ContentTemplate}" Content="{TemplateBinding Content}" ContentTransitions="{TemplateBinding ContentTransitions}" HorizontalContentAlignment="{TemplateBinding HorizontalContentAlignment}" Padding="{TemplateBinding Padding}" VerticalContentAlignment="{TemplateBinding VerticalContentAlignment}">
-                            </ContentPresenter>
-                            <animations:Implicit.Animations>
-                                <animations:ScaleAnimation Duration="0:0:0.4"></animations:ScaleAnimation>
-                            </animations:Implicit.Animations>
-                        </Grid>
-                    </ControlTemplate>
-                </Setter.Value>
-            </Setter>
-        </Style>
-    </Page.Resources>
-
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-
-        <StackPanel VerticalAlignment="Center">
-            <Button Click="Button_Click" Content="Visibility (1)"></Button>
-            <Button Click="Button_Click_1" Content="Move Me (2)"></Button>
-            <Button Click="Button_Click_2" Content="Scale Me (2)"></Button>
-        </StackPanel>
-
-        <Canvas>
-            <Border x:Name="Element" 
-                    Height="100" 
-                    Width="100" 
-                    Background="Red"
-                    Canvas.Top="100"
-                    Canvas.Left="100"
-                    extensions:VisualExtensions.CenterPoint="50,50,0">
-
-                <animations:Implicit.ShowAnimations>
-                    <animations:TranslationAnimation Duration="0:0:1" To="0, 0, 0" ></animations:TranslationAnimation>
-                    <animations:OpacityAnimation Duration="0:0:1" To="1.0"></animations:OpacityAnimation>
-                </animations:Implicit.ShowAnimations>
-
-                <animations:Implicit.HideAnimations>
-                    <animations:ScalarAnimation Target="Opacity" Duration="0:0:1" To="0.0"></animations:ScalarAnimation>
-                    <animations:ScalarAnimation Target="Translation.Y" Duration="0:0:1" To="-200">
-                        <animations:ScalarKeyFrame Key="0.1" Value="30"></animations:ScalarKeyFrame>
-                        <animations:ScalarKeyFrame Key="0.5" Value="0.0"></animations:ScalarKeyFrame>
-                    </animations:ScalarAnimation>
-                </animations:Implicit.HideAnimations>
-
-                <animations:Implicit.Animations>
-                    <animations:Vector3Animation Target="Offset"  Duration="0:0:1"></animations:Vector3Animation>
-                    <animations:ScalarAnimation Target="RotationAngleInDegrees" ImplicitTarget="Offset"  Duration="0:0:1.2" From="0" To="0">
-                        <animations:ScalarKeyFrame Key="0.9" Value="80"></animations:ScalarKeyFrame>
-                    </animations:ScalarAnimation>
-                    <animations:Vector3Animation Target="Scale" Duration="0:0:1"></animations:Vector3Animation>
-                </animations:Implicit.Animations>
-
-            </Border>
-        </Canvas>
-
-        <Button HorizontalAlignment="Right" Style="{StaticResource CustomButtonStyle}">Custom Template (No Action)</Button>
-
-    </Grid>
-</Page>
